refactor : 빌드시 에러나는 부분 수정
Ref: mental-jober#5
