--- src/main/java/com/fastcampus/jober/domain/spacewalllayout/SpaceWallLayout.java
@@ -1,49 +1,0 @@
-package com.fastcampus.jober.domain.spacewalllayout;
-
-import jakarta.persistence.*;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.Setter;
-
-import java.time.LocalDateTime;
-
-@Entity
-@Getter
-@Builder
-@Table(name = "space_wall_layout")
-public class SpaceWallLayout {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Column(length = 20, nullable = false)
-    private String name;
-
-    @Column(length = 100, nullable = false)
-    private String description;
-
-    @Column(name = "thumnail_image_url", length = 200)
-    private String thumbnailImageUrl;
-
-    @Column(length = 10, nullable = false)
-    private String type;
-
-    @Column(nullable = false)
-    private LocalDateTime createdAt;
-
-    @Column(nullable = true)
-    private LocalDateTime updatedAt;
-
-    @Column(nullable = true)
-    private LocalDateTime deletedAt;
-
-    @PrePersist
-    public void prePersist() {
-        if (this.createdAt == null) {
-            this.createdAt = LocalDateTime.now();
-        }
-    }
-
-}
-
